fix: CI clippy + stress test threshold for shared runners
- stress.rs: relax SRR 10K lookup threshold 1ms→5ms (CI shared runners)
- mount.rs: remove redundant trim() before split_whitespace()
- network.rs: fix doc_lazy_continuation (add blank line before paragraph)
- seccomp.rs: remove unnecessary i64 cast on SYS_socket
- tls_probe.rs: replace useless format!() with .to_string()

+++ crates/gvm-sandbox/src/seccomp.rs
@@ -129,7 +129,7 @@ fn build_default_filter(default_action: SeccompAction) -> Result<seccompiler::Bp
     // CLONE_NEWUSER grants apparent root + CAP_NET_ADMIN inside the user namespace,
     // but without AF_NETLINK sockets, agents cannot modify iptables rules.
     rules.insert(
-        libc::SYS_socket as i64,
+        libc::SYS_socket,
         vec![
             // Allow AF_UNIX (1) — needed for Python multiprocessing, IPC
             SeccompRule::new(vec![SeccompCondition::new(
